FINALLY fixed road occapancy bug
Added edge cases for the calculation of road occupancy for when a person
is leaving a main road to a home or office

+++ TrafficSim/Roads/IRoad.cs
@@ -6,8 +6,8 @@
         int NumberOfLanes{get; set; }
         int TimeToTraverse {get; set;}
         double Cost { get; set; }
-        ushort DirectionAOccupancy { get; set; }
-        ushort DirectionBOccupancy { get; set; }
+        int DirectionAOccupancy { get; set; }
+        int DirectionBOccupancy { get; set; }
 
     }
 }
